LLM Task Distribution System (Microservice)/6.ClaudeAI/1.Multi-LLM-Process/Workspace/llm-engine/requirements.txt
flask==2.3.3
openai==1.3.0
anthropic==0.8.1
google-cloud-aiplatform==1.36.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0